Run as unprivileged user
The application used to run as root inside the container. This
commit adds an unprivileged user, that is being used to copy the
configuration and run the executable. This hardens the application
against certain attack vectors. As far as I see, no state
should be kept in /data/ across builds, so this change shouldn't
intruduce problems.